+++ src/test/kotlin/com/github/unclepomedev/blenderprobeforpycharm/BlenderProbeUtilsTest.kt
@@ -1,5 +1,6 @@
 package com.github.unclepomedev.blenderprobeforpycharm
 
+import com.github.unclepomedev.blenderprobeforpycharm.settings.BlenderSettings
 import com.intellij.openapi.application.WriteAction
 
 class BlenderProbeUtilsTest : BaseBlenderTest() {
@@ -52,6 +53,41 @@ class BlenderProbeUtilsTest : BaseBlenderTest() {
     }
 
     fun testDetectAddonModuleName_Fallback() {
+        val expected = BlenderProbeUtils.normalizeModuleName(project.name)
+        val detectedName = BlenderProbeUtils.detectAddonModuleName(project)
+        assertEquals(expected, detectedName)
+    }
+
+    fun testDetectAddonModuleName_SettingsFallbackNameUsedWhenNoManifest() {
+        BlenderSettings.getInstance(project).state.fallbackAddonName = "legacy_addon"
+
+        val detectedName = BlenderProbeUtils.detectAddonModuleName(project)
+        assertEquals("legacy_addon", detectedName)
+    }
+
+    fun testDetectAddonModuleName_SettingsFallbackNameTrimmed() {
+        BlenderSettings.getInstance(project).state.fallbackAddonName = "  legacy_addon  "
+
+        val detectedName = BlenderProbeUtils.detectAddonModuleName(project)
+        assertEquals("legacy_addon", detectedName)
+    }
+
+    fun testDetectAddonModuleName_ManifestTakesPriorityOverSettingsFallback() {
+        BlenderSettings.getInstance(project).state.fallbackAddonName = "legacy_addon"
+
+        val baseDir = myFixture.tempDirFixture.getFile(".")!!
+        WriteAction.run<Exception> {
+            val addonDir = baseDir.createChildDirectory(this, "manifest_addon")
+            addonDir.createChildData(this, "blender_manifest.toml")
+        }
+
+        val detectedName = BlenderProbeUtils.detectAddonModuleName(project)
+        assertEquals("manifest_addon", detectedName)
+    }
+
+    fun testDetectAddonModuleName_BlankSettingsFallbackUsesProjectName() {
+        BlenderSettings.getInstance(project).state.fallbackAddonName = "   "
+
         val expected = BlenderProbeUtils.normalizeModuleName(project.name)
         val detectedName = BlenderProbeUtils.detectAddonModuleName(project)
         assertEquals(expected, detectedName)